update script tests
