+++ scripts/tests/test_dataprocessor.py
@@ -30,6 +30,32 @@ from ..machine_translation.dataprocessor import process_dataset
 from ..machine_translation.dataset import TOY
 
 
+@pytest.mark.remote_required
+def test_toy():
+    # Test toy dataset
+    train_en_de = TOY(segment='train', root='tests/data/translation_test')
+    val_en_de = TOY(segment='val', root='tests/data/translation_test')
+    test_en_de = TOY(segment='test', root='tests/data/translation_test')
+    assert len(train_en_de) == 30
+    assert len(val_en_de) == 30
+    assert len(test_en_de) == 30
+    en_vocab, de_vocab = train_en_de.src_vocab, train_en_de.tgt_vocab
+    assert len(en_vocab) == 358
+    assert len(de_vocab) == 381
+    train_de_en = TOY(segment='train', src_lang='de', tgt_lang='en',
+                      root='tests/data/translation_test')
+    de_vocab, en_vocab = train_de_en.src_vocab, train_de_en.tgt_vocab
+    assert len(en_vocab) == 358
+    assert len(de_vocab) == 381
+    for i in range(10):
+        lhs = train_en_de[i]
+        rhs = train_de_en[i]
+        assert lhs[0] == rhs[1] and rhs[0] == lhs[1]
+    time.sleep(5)
+
+
+
+
 def test_translation_preprocess():
     src_lang = 'en'
     tgt_lang = 'de'
