+++ scripts/tests/test_scripts.py
@@ -1,34 +1,9 @@
 import os
 import subprocess
+import sys
 import time
 
 import pytest
-
-from ..machine_translation.dataset import TOY
-
-
-@pytest.mark.remote_required
-def test_toy():
-    # Test toy dataset
-    train_en_de = TOY(segment='train', root='tests/data/translation_test')
-    val_en_de = TOY(segment='val', root='tests/data/translation_test')
-    test_en_de = TOY(segment='test', root='tests/data/translation_test')
-    assert len(train_en_de) == 30
-    assert len(val_en_de) == 30
-    assert len(test_en_de) == 30
-    en_vocab, de_vocab = train_en_de.src_vocab, train_en_de.tgt_vocab
-    assert len(en_vocab) == 358
-    assert len(de_vocab) == 381
-    train_de_en = TOY(segment='train', src_lang='de', tgt_lang='en',
-                      root='tests/data/translation_test')
-    de_vocab, en_vocab = train_de_en.src_vocab, train_de_en.tgt_vocab
-    assert len(en_vocab) == 358
-    assert len(de_vocab) == 381
-    for i in range(10):
-        lhs = train_en_de[i]
-        rhs = train_de_en[i]
-        assert lhs[0] == rhs[1] and rhs[0] == lhs[1]
-    time.sleep(5)
 
 
 @pytest.mark.serial
@@ -38,7 +13,7 @@ def test_toy():
 @pytest.mark.parametrize('fasttext', [True, False])
 def test_skipgram_cbow(model, fasttext):
     cmd = [
-        'python', './scripts/word_embeddings/train_sg_cbow.py', '--gpu', '0',
+        sys.executable, './scripts/word_embeddings/train_sg_cbow.py', '--gpu', '0',
         '--epochs', '2', '--model', model, '--data', 'toy', '--batch-size',
         '64']
     if fasttext:
@@ -54,7 +29,7 @@ def test_glove():
     vocab = os.path.join(path, 'word_embeddings/glove/vocab.txt')
     cooccur = os.path.join(path, 'word_embeddings/glove/cooccurrences.npz')
     cmd = [
-        'python', './scripts/word_embeddings/train_glove.py', cooccur, vocab,
+        sys.executable, './scripts/word_embeddings/train_glove.py', cooccur, vocab,
         '--batch-size', '2', '--epochs', '2']
     subprocess.check_call(cmd)
     time.sleep(5)
@@ -67,7 +42,7 @@ def test_glove():
 @pytest.mark.parametrize('fasttextloadngrams', [True, False])
 def test_embedding_evaluate_pretrained(fasttextloadngrams):
     cmd = [
-        'python', './scripts/word_embeddings/evaluate_pretrained.py',
+        sys.executable, './scripts/word_embeddings/evaluate_pretrained.py',
         '--embedding-name', 'fasttext', '--embedding-source', 'wiki.simple',
         '--gpu', '0'
     ]
@@ -88,7 +63,7 @@ def test_embedding_evaluate_from_path(evaluateanalogies, maxvocabsize):
     path = os.path.join(
         path, '../../tests/unittest/train/test_embedding/lorem_ipsum.bin')
     cmd = [
-        'python', './scripts/word_embeddings/evaluate_pretrained.py',
+        sys.executable, './scripts/word_embeddings/evaluate_pretrained.py',
         '--embedding-path', path, '--gpu', '0']
     if evaluateanalogies:
         cmd += ['--analogy-datasets', 'GoogleAnalogyTestSet']
@@ -111,14 +86,14 @@ def test_sentiment_analysis_finetune(use_pretrained):
             '--save-prefix', 'imdb_lstm_200']
     if not use_pretrained:
         args.append('--no_pretrained')
-    process = subprocess.check_call(['python', './scripts/sentiment_analysis/finetune_lm.py']+args)
+    process = subprocess.check_call([sys.executable, './scripts/sentiment_analysis/finetune_lm.py']+args)
     time.sleep(5)
 
 @pytest.mark.serial
 @pytest.mark.remote_required
 @pytest.mark.gpu
 def test_sentiment_analysis_textcnn():
-    process = subprocess.check_call(['python', './scripts/sentiment_analysis/sentiment_analysis_cnn.py',
+    process = subprocess.check_call([sys.executable, './scripts/sentiment_analysis/sentiment_analysis_cnn.py',
                                      '--gpu', '0', '--batch_size', '50', '--epochs', '1',
                                      '--dropout', '0.5', '--lr', '0.0001', '--model_mode', 'rand',
                                      '--data_name', 'MR', '--save-prefix', 'sa-model'])
@@ -133,7 +108,7 @@ def test_sampling(method):
         args.append('--use-beam-search')
     if method == 'sampling':
         args.extend(['--use-sampling', '--temperature', '1.0'])
-    process = subprocess.check_call(['python', './scripts/text_generation/sequence_sampling.py']
+    process = subprocess.check_call([sys.executable, './scripts/text_generation/sequence_sampling.py']
                                      + args)
     time.sleep(5)
 
@@ -143,7 +118,7 @@ def test_sampling(method):
 @pytest.mark.remote_required
 @pytest.mark.gpu
 def test_gnmt():
-    process = subprocess.check_call(['python', './scripts/machine_translation/train_gnmt.py', '--dataset', 'TOY',
+    process = subprocess.check_call([sys.executable, './scripts/machine_translation/train_gnmt.py', '--dataset', 'TOY',
                                      '--src_lang', 'en', '--tgt_lang', 'de', '--batch_size', '32',
                                      '--optimizer', 'adam', '--lr', '0.0025', '--save_dir', 'test',
                                      '--epochs', '1', '--gpu', '0', '--num_buckets', '5',
@@ -164,6 +139,6 @@ def test_transformer(bleu):
             '1', '--num_buckets', '5', '--bleu', bleu, '--num_units',
             '32', '--hidden_size', '64', '--num_layers', '2',
             '--num_heads', '4', '--test_batch_size', '32']
-    process = subprocess.check_call(['python', './scripts/machine_translation/train_transformer.py']
+    process = subprocess.check_call([sys.executable, './scripts/machine_translation/train_transformer.py']
                                     +args)
     time.sleep(5)
